+++ scripts/plugins/dae.mjs
@@ -12,7 +12,8 @@ export default function replaceDAESheet() {
       context = await super._preparePartContext(id, context);
       if (id === "aura") {
         context = foundry.utils.mergeObject(context, {
-          fields: this.document.system.schema.fields
+          fields: this.document.system.schema.fields,
+          isDAEEnabled: true
         }, { inplace: false });
       }
       return context;